requirements.txt
requests
numpy
streamlit
pandas
beautifulsoup4
pillow